Update about.rb
The About page does not need to be stored in an Aire Cell. To do so would mean that a copy of it would be archived with every Aire Space. As it is essentially a constant generated by the particular release of the Aire software, it should not be stored in a Space that might be accessed by different versions of the software.